update fused_moe implementation
