+++ setup.py
@@ -251,6 +251,7 @@ vllm_extension_sources = [
     "csrc/quantization/squeezellm/quant_cuda_kernel.cu",
     "csrc/quantization/gptq/q_gemm.cu",
     "csrc/cuda_utils_kernels.cu",
+    "csrc/moe_alig_block_size_kernels.cu",
     "csrc/pybind.cpp",
 ]
